back/grant-prototype/pyproject.toml
[project]
name = "grant-harness-prototype"
version = "0.1.0"
description = "AI-powered Australian government grant discovery and matching platform - Phase 1 Prototype"
authors = [{name = "Whisperlooms", email = "gordon@whisperlooms.com"}]
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.3.0", # Gemini API
    "google-cloud-aiplatform>=1.38.0", # Vertex AI (if needed)
    "crawl4ai>=0.2.0", # AI-powered web scraping
    "aiohttp>=3.9.0", # Async HTTP for crawl4ai
    "playwright>=1.40.0", # Fallback for complex sites
    "beautifulsoup4>=4.12.0", # HTML parsing
    "lxml>=4.9.0", # XML/HTML processing
    "requests>=2.31.0", # HTTP client
    "pydantic>=2.5.0", # Data validation
    "python-dotenv>=1.0.0", # Environment management
    "python-dateutil>=2.8.0", # Australian date parsing
    "pandas>=2.1.0", # Data manipulation
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.11.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "ipython>=8.17.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scrapers", "gemini_store", "matching", "models"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=scrapers --cov=gemini_store --cov=matching --cov=models --cov-report=term-missing"
